devtools/client/shared/components/test/mochitest/test_reps_window.html


<!DOCTYPE HTML>
<html>
<!--
Test window rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep tests - window</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    let React = browserRequire("devtools/client/shared/vendor/react");
    let { Rep } = browserRequire("devtools/client/shared/components/reps/rep");
    let { Window } = browserRequire("devtools/client/shared/components/reps/window");

    let gripStub = {
      "type": "object",
      "class": "Window",
      "actor": "server1.conn3.obj198",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 887,
      "preview": {
        "kind": "ObjectWithURL",
        "url": "about:newtab"
      }
    };

    // Test that correct rep is chosen
    const renderedRep = shallowRenderComponent(Rep, { object: gripStub });
    is(renderedRep.type, Window.rep, `Rep correctly selects ${Window.rep.displayName}`);

    // Test rendering
    const renderedComponent = renderComponent(Window.rep, { object: gripStub });
    ok(renderedComponent.className.includes("objectBox-Window"), "Window rep has expected class name");
    const innerNode = renderedComponent.querySelector(".objectPropValue");
    is(innerNode.textContent, "about:newtab", "Window rep has expected inner HTML structure and text content");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
